gui_pysimplegui/requirements.txt
python-vlc
PySimpleGUI